st/mesa: add int cast to silence warning
MSVC warns that negating an unsigned value yields an unsigned value.

+++ src/mesa/state_tracker/st_cb_fbo.c
@@ -750,7 +750,7 @@ st_MapRenderbuffer(struct gl_context *ctx,
                             usage, x, y2, w, h, &strb->transfer);
    if (map) {
       if (invert) {
-         *rowStrideOut = -strb->transfer->stride;
+         *rowStrideOut = -(int) strb->transfer->stride;
          map += (h - 1) * strb->transfer->stride;
       }
       else {
